refactor: 사용 메서드 위치 변경 import 문 최적화
